+++ ydb/core/kqp/compute_actor/kqp_scan_compute_actor.h
@@ -3,7 +3,7 @@
 #include "kqp_scan_events.h"
 
 #include <ydb/core/kqp/runtime/kqp_scan_data.h>
-#include <ydb/core/kqp/runtime/scheduler/kqp_compute_scheduler.h>
+#include <ydb/core/kqp/runtime/scheduler/kqp_schedulable_actor.h>
 #include <ydb/library/yql/dq/actors/compute/dq_compute_actor_async_io.h>
 #include <ydb/library/yql/dq/actors/compute/dq_compute_actor.h>
