+++ ydb/core/kqp/runtime/scheduler/ya.make
@@ -9,6 +9,7 @@ PEERDIR(
 SRCS(
     kqp_compute_pool.cpp
     kqp_compute_scheduler.cpp
+    kqp_schedulable_actor.cpp
 )
 
 YQL_LAST_ABI_VERSION()
